+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Fix crashes caused by missing permissions when checking for Steam apps
+
 ## [1.10.1]- 2022-12-10
 ### Fixed
  - Fix crash when unknown XDG Flatpak filesystem permissions are enabled