+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.10.2] - 2023-02-13
 ### Added
  - Launch application with fixed locale settings if Steam Deck is used and non-existent locales are configured
 
@@ -248,7 +248,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.10.1...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.10.2...HEAD
+[1.10.2]: https://github.com/Matoking/protontricks/compare/1.10.1...1.10.2
 [1.10.1]: https://github.com/Matoking/protontricks/compare/1.10.0...1.10.1
 [1.10.0]: https://github.com/Matoking/protontricks/compare/1.9.2...1.10.0
 [1.9.2]: https://github.com/Matoking/protontricks/compare/1.9.1...1.9.2